backend/requirements.txt
httpx==0.24.1
daphne==3.0.2
adrf==0.1.1
python-bioformats==4.0.7
Django==4.2.1
django-cors-headers==4.0.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
opencv-python==4.7.0.72